WebHost: offer API to modify WebHost
