+++ WebHostLib/tracker.py
@@ -1447,21 +1447,6 @@ def get_multiworld_tracker(tracker: UUID):
 
     return render_template("multiTracker.html", **data)
 
-if "Factorio" in games:
-    @app.route('/tracker/<suuid:tracker>/Factorio')
-    @cache.memoize(timeout=60)  # multisave is currently created at most every minute
-    def get_Factorio_multiworld_tracker(tracker: UUID):
-        data = _get_multiworld_tracker_data(tracker)
-        if not data:
-            abort(404)
-
-        data["inventory"] = _get_inventory_data(data)
-        data["enabled_multiworld_trackers"] = get_enabled_multiworld_trackers(data["room"], "Factorio")
-        data["item_name_to_id"] =  games["Factorio"]["location_name_to_id"]
-
-        return render_template("multiFactorioTracker.html", **data)
-
-
 @app.route('/tracker/<suuid:tracker>/A Link to the Past')
 @cache.memoize(timeout=60)  # multisave is currently created at most every minute
 def get_LttP_multiworld_tracker(tracker: UUID):
@@ -1590,6 +1575,3 @@ game_specific_trackers: typing.Dict[str, typing.Callable] = {
 multi_trackers: typing.Dict[str, typing.Callable] = {
     "A Link to the Past": get_LttP_multiworld_tracker,
 }
-
-if "Factorio" in games:
-    multi_trackers["Factorio"] = get_Factorio_multiworld_tracker
