+++ worlds/alttp/__init__.py
@@ -807,7 +807,15 @@ class ALTTPWorld(World):
                 }
             )
         return slot_data
-        
+
+    @classmethod
+    def run_webhost_setup(cls):
+        rom_file = get_base_rom_path()
+        if os.path.exists(rom_file):
+            from .Sprites import update_sprites
+            update_sprites()
+        else:
+            logging.warning("Could not update LttP sprites.")
 
 def get_same_seed(world, seed_def: tuple) -> str:
     seeds: typing.Dict[tuple, str] = getattr(world, "__named_seeds", {})
