+++ worlds/factorio/__init__.py
@@ -524,6 +524,32 @@ class Factorio(World):
                             all_items[name], self.player)
         return item
 
+    @classmethod
+    def run_webhost_app_setup(cls, app):
+        from uuid import UUID
+
+        from werkzeug.exceptions import abort
+        from flask import render_template
+
+        from WebHostLib import cache
+        from WebHostLib.tracker import (_get_multiworld_tracker_data, _get_inventory_data,
+                                        get_enabled_multiworld_trackers, multi_trackers)
+
+
+        @app.route('/tracker/<suuid:tracker>/Factorio')
+        @cache.memoize(timeout=60)  # multisave is currently created up to every minute
+        def get_Factorio_multiworld_tracker(tracker: UUID):
+            data = _get_multiworld_tracker_data(tracker)
+            if not data:
+                abort(404)
+
+            data["inventory"] = _get_inventory_data(data)
+            data["enabled_multiworld_trackers"] = get_enabled_multiworld_trackers(data["room"], "Factorio")
+            data["item_name_to_id"] = cls.location_name_to_id
+
+            return render_template("multiFactorioTracker.html", **data)
+
+        multi_trackers[cls.game] = get_Factorio_multiworld_tracker
 
 class FactorioLocation(Location):
     game: str = Factorio.game
